+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make the chapter icon in the generated PDF table of contents scale with the configured font size.
 - Improve line wrapping in generated PDF code blocks by adding more break opportunities and scaling continuation indentation with the configured code font size.
 
+### Fixed
+- Fix an issue where `_?_` in grammar notes could be rendered incorrectly in generated PDF output.
+
 ## [2.3.1] - 2026-04-04
 ### Fixed
 - Fix an issue where `Summary of the Grammar` generation could fall out of sync with the latest upstream `swift-book` repo after the summary build logic moved from `bin/publish-book` to `bin/generate-grammar`.